重构删除文章

+++ app/models/article/article.go
@@ -6,11 +6,11 @@ import (
 )
 
 type Article struct {
-	ID    int
+	ID    int64
 	Title string
 	Body  string
 }
 
 func (a Article) Link() string {
-	return route.Name2URL("articles.show", "id", strconv.FormatInt(int64(a.ID), 10))
+	return route.Name2URL("articles.show", "id", strconv.FormatInt(a.ID, 10))
 }
